feat: storybook Counter의 숫자 업데이트를 위한 state, updater 추가
